move tests to test/support/tests.mjs
